Add entity I/O inputs to mediaplayer_base entity
`AddPlayer` adds the activator to the list of media listeners.
`RemovePlayer` removes the activator from the list of media listeners.
`PlayPauseMedia` toggles the play or pause state of the current media.
`SkipMedia` skips the current media.

Shoutout to the Garry's Mod Jazztronauts gamemode which will be making
use of this. https://twitter.com/JazzSourceMod